add motors

+++ src/rc.h
@@ -3,6 +3,6 @@
 #include <Arduino.h>
 
 void setup_pwmRead();
- boolean RC_avail();
- void print_RCpwm();
-   float RC_decode(int CH);
+boolean RC_avail();
+void print_RCpwm();
+int RC_decode(int CH);
